requirements.txt
asgiref==3.5.2
Django==4.1.4
django-autoslug==1.9.9
django-ckeditor==6.5.1
django-js-asset==2.0.0
gunicorn==20.1.0
Pillow==9.5.0
psycopg2-binary==2.9.6
serpent==1.41
sqlparse==0.4.3
tzdata==2022.7
tcod==15.0.1
tdl==6.0.0
numpy==1.24.2
typing_extensions==4.5.0